requirements.txt
ccxt==3.0.86
pandas==2.1.1
pandas_ta==0.3.21b0
python-telegram-bot==13.21